refactor: Extract updateTile from updateMap
RULE
IF ONLY AT THE START

SMELL
Functions doing more
than one thing

STATEMENT
If you have an if, it should be the first thing in the function.

EXPLANATION
if a function has an if, it should be the first thing in the func-
tion. It should also be the only thing, in the sense that we should not do anything after
it