refactor: Extract handleInput from handleInputs
RULE
IF ONLY AT THE START

SMELL
Functions doing more
than one thing

STATEMENT
If you have an if, it should be the first thing in the function.

EXPLANATION
if a function has an if, it should be the first thing in the function.
It should also be the only