refactor: Replace Input enum with Input interface
This commit takes the first step in the refactoring process to remove the
"if-else" structure as per the rule "NEVER USE IF WITH ELSE."

REPLACE TYPE CODE WITH CLASSES
This refactoring pattern transforms an enum into an interface, and the
enums’ values become classes. Doing so enables us to add properties
to each value and localize functionality concerning that specific value.